pyproject.toml
[tool.poetry]
name = "birdclef2025"
version = "0.1.0"
description = "Packages for BirdCLEF2025"
authors = ["Ivan Bashtovyi"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
torch = "^2.6.0"
torchaudio = "^2.6.0"
audiomentations = "^0.40.0"
pandas = "^2.2.3"
timm = "^1.0.15"
lightning = "^2.5.1"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
matplotlib = "<3.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
